requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==4.1.0
build==1.2.1
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
constable==0.3.1
coverage==7.5.4
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.8
django==5.0.6
django-allauth==0.63.3
django-appconf==1.0.6
django-classy-tags==4.1.0
django-compressor==4.5
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-formtools==2.5.1
django-fsm==2.8.2
django-imagekit==5.0.0
django-linear-migrations==2.13.0
django-modelcluster==6.3
django-parler==2.3
django-permissionedforms==0.1
django-phonenumber-field==8.0.0
django-polymorphic==3.1.0
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
factory-boy==3.3.0
faker==25.9.1
filelock==3.15.4
filetype==1.2.0
html5lib==1.1
identify==2.5.36
idna==3.7
iniconfig==2.0.0
l18n==2021.3
laces==0.1.1
lxml==5.2.2
markdown==3.6
nodeenv==1.9.1
oauthlib==3.2.2
openpyxl==3.1.4
packaging==24.1
phonenumberslite==8.13.40
pilkit==3.0
pillow==10.3.0
pillow-heif==0.16.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycparser==2.22
pygments==2.18.0
pyjwt==2.8.0
pypng==0.20220715.0
pyproject-hooks==1.1.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
pyyaml==6.0.1
qrcode==7.4.2
rcssmin==1.1.2
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
setuptools==70.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
svglib==1.5.1
telepath==0.3.1
tinycss2==1.3.0
typing-extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wagtail==6.1.2
wagtail-markdown==0.12.1
webencodings==0.5.1
willow==1.8.0
